DDCNL-3750 adding test cases to HttpHandlerSpec
