+++ app/views/employments/update/WhatDoYouWantToTellUsView.scala.html
@@ -25,7 +25,7 @@
 
 @(
     viewmodel: EmploymentViewModel,
-    updateEmploymentDetailsForm: Form[String],
+    updateEmploymentDetailsForm: Form[String]
 )( implicit request: Request[_],
     user: controllers.auth.AuthedUser,
     messages: Messages,
