+++ app/views/incomes/CalcUnavailableView.scala.html
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *@
+
 @import scala.concurrent.ExecutionContext
 
 @this(mainTemplate: MainTemplate)
